Issue #1302 Merge branch '5-Data_Delivery' into development
Conflicts:
    edexOsgi/build.edex/deploy-common/external-rules.xml
    edexOsgi/build.edex/deploy-common/plugin-methods.xml
    edexOsgi/com.raytheon.edex.plugin.text/src/com/raytheon/edex/plugin/text/dao/TextDao.java
    edexOsgi/com.raytheon.uf.edex.database/src/com/raytheon/uf/edex/database/plugin/PluginDao.java
    edexOsgi/com.raytheon.uf.edex.esb.camel/src/com/raytheon/uf/edex/esb/camel/Executor.java
    edexOsgi/build.edex/build.xml
    edexOsgi/build.edex/esb/conf/modes.xml
    edexOsgi/com.raytheon.edex.feature.uframe/feature.xml
    edexOsgi/com.raytheon.edex.plugin.grib/res/spring/grib-ingest.xml
    edexOsgi/com.raytheon.edex.plugin.grib/src/com/raytheon/edex/plugin/grib/Grib1Decoder.java
    edexOsgi/com.raytheon.edex.plugin.grib/src/com/raytheon/edex/plugin/grib/GribDecoder.java
    edexOsgi/com.raytheon.uf.common.gridcoverage/src/com/raytheon/uf/common/gridcoverage/GridCoverage.java


Former-commit-id: abecf8df23104748e611e25e5cfcdf0b0ffb3a1e [formerly abecf8df23104748e611e25e5cfcdf0b0ffb3a1e [formerly 8e69107b5afe12539b5fcc9ac6039b704951216f]] [formerly 98d7beae5064edc8a52f374a4fffb6c98b5c2585] [formerly abecf8df23104748e611e25e5cfcdf0b0ffb3a1e [formerly abecf8df23104748e611e25e5cfcdf0b0ffb3a1e [formerly 8e69107b5afe12539b5fcc9ac6039b704951216f]] [formerly 98d7beae5064edc8a52f374a4fffb6c98b5c2585] [formerly bcf575966b550519e01a35bd8091f358b5d36800 [formerly bcf575966b550519e01a35bd8091f358b5d36800 [formerly 31d7d09c10c841cfd0ac019699249840668963bc]] [formerly 98d7beae5064edc8a52f374a4fffb6c98b5c2585 [formerly 98d7beae5064edc8a52f374a4fffb6c98b5c2585 [formerly 89583d97c8823818f87dc1a234078d85ba7bea4a]]]]]
Former-commit-id: bcf575966b550519e01a35bd8091f358b5d36800 [formerly bcf575966b550519e01a35bd8091f358b5d36800 [formerly 31d7d09c10c841cfd0ac019699249840668963bc]]
Former-commit-id: a7611084a1fb99eae4a7cff1983deded1e8541ba [formerly 5e93f5347e4c8a04d4c5a5bbfd98c80f4f398a53] [formerly aa043fdafa5f83bf035ed570413fe0501f8ca8c2 [formerly e7a0f96a9a12671a5090f9ecae8a4d6332929ed7]] [formerly bcf575966b550519e01a35bd8091f358b5d36800]
Former-commit-id: 1265fb8e10123fceecacec44ef3139e8bb593389 [formerly a733d0e7e6a6ee56de8bda042870b76037a8834f] [formerly ba5e9428ecf831d5888e819deab74f1989c375bd [formerly 4d9c9453883002ac4d64e28f05cf168b4fa4d838]]
Former-commit-id: 888b87cf333197cb1e7178aae73c65da1c676f53 [formerly 931d94d1def4fe6811aa7ab85f8e4710202c1862]
Former-commit-id: 6ef881da26113493caf0042038104a23f2657a00

+++ edexOsgi/com.raytheon.edex.plugin.airep/unit-test/test/airep/TestAIREPParser.java
@@ -267,20 +267,20 @@ public class TestAIREPParser {
         // Winds with "KT"
         p = new AirepParser("ARP DAL278 3336N 165E 1543 F320 MS40 110/010KT TB LGT=", c);
         assertNotNull(p);
-        assertEquals(110, p.getWindDirection());
-        assertEquals(10, p.getWindSpeed());
+        assertEquals(110, p.getWindDirection().intValue());
+        assertEquals(10, p.getWindSpeed().intValue());
 
         // Winds with "KTS"
         p = new AirepParser("ARP DAL278 3336N 165E 1543 F320 MS40 265/010KTS TB LGT=", c);
         assertNotNull(p);
-        assertEquals(265, p.getWindDirection());
-        assertEquals(10, p.getWindSpeed());
+        assertEquals(265, p.getWindDirection().intValue());
+        assertEquals(10, p.getWindSpeed().intValue());
 
         // Winds with no units - assume knots
         p = new AirepParser("ARP DAL278 3336N 165E 1543 F320 MS40 265/010 TB LGT=", c);
         assertNotNull(p);
-        assertEquals(265, p.getWindDirection());
-        assertEquals(10, p.getWindSpeed());
+        assertEquals(265, p.getWindDirection().intValue());
+        assertEquals(10, p.getWindSpeed().intValue());
         
     }
